refactor: extract excel range helpers
